Version 0.4: Font rendering & Text layout, User Interface basis, Tracy Profiler & GPU instrumentation, Memory and Performance optimizations

+++ Build/Posix/Build.sh
@@ -20,22 +20,32 @@ case "${UNAME_OUT}" in
 esac
 
 BUILD_TYPE=Release
-BUILD_VERSION=0.2
+BUILD_VERSION=0.4
 
 SCRIPT_DIR=$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )
 SOURCE_DIR=$SCRIPT_DIR/../..
 OUTPUT_DIR=$SCRIPT_DIR/../Output/$CMAKE_GENERATOR
 INSTALL_DIR=$OUTPUT_DIR/Install
-CMAKE_FLAGS="-DMETHANE_VERSION=$BUILD_VERSION \
+CMAKE_FLAGS=" \
     -DMETHANE_SHADERS_CODEVIEW_ENABLED:BOOL=ON \
-    -DMETHANE_ITT_INSTRUMENTATION_ENABLED:BOOL=ON \
-    -DMETHANE_SCOPE_TIMERS_ENABLED:BOOL=OFF \
     -DMETHANE_RUN_TESTS_DURING_BUILD:BOOL=OFF \
+    -DMETHANE_COMMAND_DEBUG_GROUPS_ENABLED:BOOL=ON \
+    -DMETHANE_LOGGING_ENABLED:BOOL=OFF \
     -DMETHANE_USE_OPEN_IMAGE_IO:BOOL=OFF \
-    -DMETHANE_COMMAND_EXECUTION_LOGGING:BOOL=OFF"
+    -DMETHANE_SCOPE_TIMERS_ENABLED:BOOL=OFF \
+    -DMETHANE_ITT_INSTRUMENTATION_ENABLED:BOOL=ON \
+    -DMETHANE_ITT_METADATA_ENABLED:BOOL=OFF \
+    -DMETHANE_GPU_INSTRUMENTATION_ENABLED:BOOL=OFF \
+    -DMETHANE_TRACY_PROFILING_ENABLED:BOOL=OFF \
+    -DMETHANE_TRACY_PROFILING_ON_DEMAND:BOOL=OFF"
+
+GRAPHVIZ_DIR=$OUTPUT_DIR/GraphViz
+GRAPHVIZ_DOT_DIR=$GRAPHVIZ_DIR/dot
+GRAPHVIZ_IMG_DIR=$GRAPHVIZ_DIR/img
+GRAPHVIZ_FILE=MethaneKit.dot
+GRAPHVIZ_DOT_EXE=$(which dot)
 
 if [ "$IS_ANALYZE_BUILD" == true ]; then
-
     BUILD_DIR=$OUTPUT_DIR/Analyze
     SONAR_SCANNER_DIR=$SOURCE_DIR/Externals/SonarScanner/binaries/MacOS
     SONAR_SCANNER_ZIP=$SONAR_SCANNER_DIR.zip
@@ -58,26 +68,30 @@ else
 fi
 
 rm -rf "$OUTPUT_DIR"
-mkdir -p "$BUILD_DIR"
-
-echo Pulling latest changes from submodules...
-git submodule update --init --recursive
-
-echo ---
+if ! mkdir -p "$BUILD_DIR"; then
+    echo "Failed to create clean build directory."
+    exit 1
+fi
 
 if [ "$IS_ANALYZE_BUILD" == true ]; then
 
     echo Unpacking Sonar Scanner binaries...
     rm -rf "$SONAR_SCANNER_DIR"
-    unzip -q "$SONAR_SCANNER_ZIP" -d "$SONAR_SCANNER_DIR"
+    if ! unzip -q "$SONAR_SCANNER_ZIP" -d "$SONAR_SCANNER_DIR"; then
+        echo "Sonar-scanner binaries unpack failed."
+        exit 1
+    fi
 
     GITBRANCH=$(git symbolic-ref --short HEAD)
 
     echo Analyzing code with Sonar Scanner on branch $GITBRANCH...
-    "$SONAR_BUILD_WRAPPER_EXE" --out-dir "$BUILD_DIR" \
-        cmake -H"$SOURCE_DIR" -B"$BUILD_DIR" -G "$CMAKE_GENERATOR" $CMAKE_FLAGS
+    if ! "$SONAR_BUILD_WRAPPER_EXE" --out-dir "$BUILD_DIR" \
+        cmake -H"$SOURCE_DIR" -B"$BUILD_DIR" -G "$CMAKE_GENERATOR" $CMAKE_FLAGS; then
+        echo "Sonar-scanner CMake generation failed."
+        exit 1
+    fi
 
-    "$SONAR_SCANNER_EXE" \
+    if ! "$SONAR_SCANNER_EXE" \
         -Dsonar.projectKey=egorodet_MethaneKit \
         -Dsonar.organization=egorodet-github \
         -Dsonar.branch.name=$GITBRANCH \
@@ -85,13 +99,50 @@ if [ "$IS_ANALYZE_BUILD" == true ]; then
         -Dsonar.sources="$SOURCE_DIR" \
         -Dsonar.cfamily.build-wrapper-output="$BUILD_DIR" \
         -Dsonar.host.url=https://sonarcloud.io \
-        -Dsonar.login=6e1dbce6af614f59d75f1d78f0609aaaa60caee1
+        -Dsonar.login=6e1dbce6af614f59d75f1d78f0609aaaa60caee1; then
+        echo "Sonar-scanner build failed."
+        exit 1
+    fi
 
 else
-    echo Building with $CMAKE_GENERATOR...
-    cmake -H"$SOURCE_DIR" -B"$BUILD_DIR" -G "$CMAKE_GENERATOR" -DCMAKE_INSTALL_PREFIX="$INSTALL_DIR" $CMAKE_FLAGS
-    cmake --build "$BUILD_DIR" --config $BUILD_TYPE --target install
 
+    echo ----------
+    echo Generating build files for $CMAKE_GENERATOR...
+    if ! cmake -S "$SOURCE_DIR" -B "$BUILD_DIR" -G "$CMAKE_GENERATOR" \
+        --graphviz="$GRAPHVIZ_DOT_DIR/$GRAPHVIZ_FILE" \
+        -DCMAKE_INSTALL_PREFIX="$INSTALL_DIR" $CMAKE_FLAGS; then
+        echo "Methane CMake generation failed."
+        exit 1
+    fi
+
+    echo ----------
+    if [ -x "$GRAPHVIZ_DOT_EXE" ]; then
+        echo Converting GraphViz diagrams to images...
+        mkdir "$GRAPHVIZ_IMG_DIR"
+        for DOT_PATH in $GRAPHVIZ_DOT_DIR/*; do
+            DOT_IMG=$GRAPHVIZ_IMG_DIR/${DOT_PATH##*/}.png
+            echo Writing image $DOT_IMG...
+            if ! "$GRAPHVIZ_DOT_EXE" -Tpng "$DOT_PATH" -o "$DOT_IMG"; then
+                echo "Dot failed to generate diagram image."
+                exit 1
+            fi
+        done
+    else
+        echo GraphViz dot executable was not found. Skipping graph images generation.
+    fi
+
+    echo ----------
+    echo Build with $CMAKE_GENERATOR...
+    if ! cmake --build "$BUILD_DIR" --config $BUILD_TYPE --target install; then
+        echo "Methane build failed."
+        exit 1
+    fi
+
+    echo ----------
     echo Running unit-tests...
-    cmake -E chdir "$BUILD_DIR" ctest --build-config $BUILD_TYPE --output-on-failure
+    if ! cmake -E chdir "$BUILD_DIR" ctest --build-config $BUILD_TYPE --output-on-failure; then
+        echo "Methane tests failed."
+        exit 1
+    fi
+
 fi
